requirements.txt
uvicorn[standard]==0.27.1
fastapi==0.110.0
pydantic==2.6.4
sqlalchemy==2.0.29
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pymysql==1.1.0
python-multipart
alembic
python-dotenv
pydantic[email]
